Handle 32bits & 64bits binary

+++ nm/libft/ft_intlen_base.c
@@ -6,13 +6,13 @@
 /*   By: julien <julien@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/17 15:08:45 by jballang          #+#    #+#             */
-/*   Updated: 2018/06/22 11:07:11 by julien           ###   ########.fr       */
+/*   Updated: 2018/06/29 15:24:41 by julien           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_intlen_base(long long int n, int base)
+int	ft_intlen_base(int n, int base)
 {
 	int len;
